refactor: split docs public models
