+++ crates/ox_content_docs/src/graph.rs
@@ -11,256 +11,26 @@ use oxc_parser::Parser;
 use oxc_resolver::{ResolveOptions, Resolver, TsconfigOptions, TsconfigReferences};
 use oxc_span::SourceType;
 use rustc_hash::{FxBuildHasher, FxHashMap, FxHashSet};
-use serde::{Deserialize, Serialize};
-use thiserror::Error;
+
+mod error;
+mod model;
+mod options;
+
+pub use error::GraphError;
+pub use model::{
+    DocsDiagnostic, DocsDiagnosticCode, EntrypointDocsModule, EntrypointModule, ExportGraph,
+    ExportKind, ExportSource, PublicExport, ResolvedModule,
+};
+pub use options::{
+    EntryPointDocsOptions, EntryPointSpec, ExternalDocsOptions, ExternalPackageSource, GraphOptions,
+};
 
 #[allow(unused_imports)]
 use crate::profile_span;
 use crate::string_builder::{join2, join4, StringBuilder};
 use crate::{
-    normalize_doc_items, ApiDocTag, DocExtractor, DocItem, ExtractError, NormalizedDocEntry,
-    NormalizedDocKind,
+    normalize_doc_items, ApiDocTag, DocExtractor, DocItem, NormalizedDocEntry, NormalizedDocKind,
 };
-
-/// Entry point used to group generated API docs.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct EntryPointSpec {
-    /// Source file path.
-    pub path: PathBuf,
-    /// Public module name. Defaults to the file stem.
-    pub name: Option<String>,
-}
-
-/// Export graph resolution options.
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct GraphOptions {
-    /// Root used to resolve relative entry point paths.
-    pub root: Option<PathBuf>,
-    /// Optional TypeScript config for path alias resolution.
-    pub tsconfig: Option<PathBuf>,
-    /// External package documentation extraction options.
-    #[serde(default)]
-    pub external_docs: ExternalDocsOptions,
-}
-
-/// External package documentation extraction options.
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ExternalDocsOptions {
-    /// Resolve external package re-exports into documentation entries.
-    pub enabled: bool,
-    /// Explicit package source entries used before package exports/types resolution.
-    #[serde(default)]
-    pub package_sources: Vec<ExternalPackageSource>,
-}
-
-/// Explicit source entry for an external package.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ExternalPackageSource {
-    /// Package name or exact package subpath specifier.
-    pub package: String,
-    /// Source or declaration entry file.
-    pub entry: PathBuf,
-}
-
-/// Options for extracting docs grouped by entry point.
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct EntryPointDocsOptions {
-    /// Export graph options.
-    pub graph: GraphOptions,
-    /// Include `@private` docs.
-    pub include_private: bool,
-    /// Include `@internal` docs.
-    pub include_internal: bool,
-    /// Opt in to TSDoc-style type-parameter docs (`@typeParam` / `<T>` table).
-    /// Off by default (JSDoc semantics).
-    #[serde(default)]
-    pub type_parameters: bool,
-}
-
-/// Resolved export graph.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ExportGraph {
-    /// Public entry point modules.
-    pub entrypoints: Vec<EntrypointModule>,
-    /// Resolved modules keyed by absolute source path.
-    pub modules: FxHashMap<PathBuf, ResolvedModule>,
-}
-
-/// Public entry point module.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct EntrypointModule {
-    /// Public module name.
-    pub name: String,
-    /// Source file path.
-    pub source_path: PathBuf,
-    /// Public exports reachable from this entry point.
-    pub exports: Vec<PublicExport>,
-}
-
-/// Resolved source module.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ResolvedModule {
-    /// Source file path.
-    pub path: PathBuf,
-    /// Exports declared or re-exported by this module.
-    pub exports: Vec<PublicExport>,
-}
-
-/// Public export metadata.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct PublicExport {
-    /// Public export name after aliasing.
-    pub name: String,
-    /// Export kind.
-    pub kind: ExportKind,
-    /// Export source.
-    pub source: ExportSource,
-}
-
-/// Export kind.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub enum ExportKind {
-    /// Runtime value export.
-    Value,
-    /// Type-only export.
-    Type,
-    /// Export available as both value and type.
-    ValueAndType,
-    /// Namespace export.
-    Namespace,
-    /// Default export.
-    Default,
-}
-
-/// Export source.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "kind", rename_all = "camelCase")]
-pub enum ExportSource {
-    /// Export resolved to a local source file.
-    Local {
-        /// Source module path.
-        module: PathBuf,
-        /// Original exported name in the source module.
-        original_name: String,
-    },
-    /// Export from an external package.
-    External {
-        /// Package name.
-        package: String,
-        /// Original module specifier.
-        #[serde(default, skip_serializing_if = "String::is_empty")]
-        specifier: String,
-        /// Resolved source or declaration module path, when available.
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        module: Option<PathBuf>,
-        /// Original exported name.
-        original_name: String,
-        /// Whether the export is type-only.
-        type_only: bool,
-    },
-}
-
-/// Docs grouped by a public entry point.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct EntrypointDocsModule {
-    /// Public module name.
-    pub name: String,
-    /// Virtual docs module file used by Markdown generation.
-    pub file: String,
-    /// Source file path.
-    pub source_path: PathBuf,
-    /// Module-level description from the entry file's `@module` block or leading
-    /// file comment. Empty when the entry file has no module-level JSDoc.
-    #[serde(default)]
-    pub description: String,
-    /// Module-level example blocks from the entry file's `@module` block or
-    /// leading file comment.
-    #[serde(default)]
-    pub examples: Vec<String>,
-    /// Module-level custom JSDoc tags.
-    #[serde(default)]
-    pub tags: Vec<ApiDocTag>,
-    /// Normalized docs entries for reachable exports.
-    pub entries: Vec<NormalizedDocEntry>,
-    /// Public export metadata, including external re-exports.
-    pub exports: Vec<PublicExport>,
-    /// Diagnostics for exports that could not be emitted as docs entries.
-    #[serde(default)]
-    pub diagnostics: Vec<DocsDiagnostic>,
-}
-
-/// Diagnostic for an entry point export during docs extraction.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct DocsDiagnostic {
-    /// Machine-readable diagnostic code.
-    pub code: DocsDiagnosticCode,
-    /// Public entry point name.
-    pub entrypoint: String,
-    /// Public export name.
-    pub export_name: String,
-    /// Public export kind.
-    pub export_kind: ExportKind,
-    /// Export source metadata.
-    pub source: ExportSource,
-    /// Human-readable diagnostic message.
-    pub message: String,
-}
-
-/// Diagnostic code for entry point docs extraction.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub enum DocsDiagnosticCode {
-    /// Export was intentionally filtered by visibility options.
-    FilteredByVisibility,
-    /// Graph export could not be matched to a declaration.
-    MissingDeclaration,
-    /// Export kind is not emitted as a docs entry.
-    UnsupportedExport,
-    /// External export could not be resolved to a source or declaration module.
-    UnresolvedExternal,
-}
-
-/// Export graph error.
-#[derive(Debug, Error)]
-pub enum GraphError {
-    /// IO error while reading a module.
-    #[error("failed to read {path}: {source}")]
-    Read {
-        /// Path that failed to read.
-        path: PathBuf,
-        /// Source error.
-        #[source]
-        source: std::io::Error,
-    },
-    /// Parser error.
-    #[error("failed to parse {path}: {message}")]
-    Parse {
-        /// Path that failed to parse.
-        path: PathBuf,
-        /// Parser message.
-        message: String,
-    },
-    /// Resolver error.
-    #[error("failed to resolve {specifier} from {importer}: {message}")]
-    Resolve {
-        /// Importer path.
-        importer: PathBuf,
-        /// Module specifier.
-        specifier: String,
-        /// Resolver message.
-        message: String,
-    },
-    /// Documentation extraction error.
-    #[error("failed to extract docs from {path}: {source}")]
-    Extract {
-        /// Path that failed to extract.
-        path: PathBuf,
-        /// Source error.
-        #[source]
-        source: ExtractError,
-    },
-}
 
 /// Builds an export graph from entry points.
 ///
