+++ crates/ox_content_docs/src/normalize.rs
@@ -4,9 +4,20 @@
 use std::collections::BTreeMap;
 
 use phf::phf_set;
-use serde::{Deserialize, Serialize};
 
-use crate::extractor::{DocItem, DocItemKind, DocTag, ParamDoc, TypeParamDoc};
+mod entry;
+mod kind;
+mod member;
+mod model;
+
+pub use entry::NormalizedDocEntry;
+pub use kind::{NormalizedDocKind, NormalizedMemberKind};
+pub use member::NormalizedMember;
+pub use model::{
+    NormalizedParamDoc, NormalizedReturnDoc, NormalizedThrowsDoc, NormalizedTypeParam,
+};
+
+use crate::extractor::{DocItem, DocTag, ParamDoc, TypeParamDoc};
 #[allow(unused_imports)]
 use crate::profile_span;
 
@@ -35,325 +46,6 @@ static TYPE_PARAM_TAG_NAMES: phf::Set<&'static str> = phf_set! {
     "typeParam",
     "template",
 };
-
-/// Documentation item kind supported by the generated API reference.
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum NormalizedDocKind {
-    /// Function declaration or function-valued variable.
-    #[default]
-    Function,
-    /// Class declaration.
-    Class,
-    /// TypeScript interface declaration.
-    Interface,
-    /// Type alias.
-    Type,
-    /// Enum declaration.
-    Enum,
-    /// Variable declaration.
-    Variable,
-    /// Module or namespace.
-    Module,
-}
-
-impl NormalizedDocKind {
-    /// Converts extractor-level kinds into public documentation kinds.
-    #[must_use]
-    pub fn from_doc_item_kind(kind: DocItemKind) -> Option<Self> {
-        match kind {
-            DocItemKind::Function => Some(Self::Function),
-            DocItemKind::Class => Some(Self::Class),
-            DocItemKind::Interface => Some(Self::Interface),
-            DocItemKind::Type => Some(Self::Type),
-            DocItemKind::Enum => Some(Self::Enum),
-            DocItemKind::Variable => Some(Self::Variable),
-            DocItemKind::Module => Some(Self::Module),
-            DocItemKind::Method
-            | DocItemKind::Property
-            | DocItemKind::Constructor
-            | DocItemKind::Getter
-            | DocItemKind::Setter
-            | DocItemKind::EnumMember
-            | DocItemKind::IndexSignature => None,
-        }
-    }
-
-    /// Returns the JavaScript-facing string for this kind.
-    #[must_use]
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::Function => "function",
-            Self::Class => "class",
-            Self::Interface => "interface",
-            Self::Type => "type",
-            Self::Enum => "enum",
-            Self::Variable => "variable",
-            Self::Module => "module",
-        }
-    }
-}
-
-/// Documentation item kind supported for class/interface/type/enum members.
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum NormalizedMemberKind {
-    /// Object or class property.
-    #[default]
-    Property,
-    /// Method signature.
-    Method,
-    /// Class constructor.
-    Constructor,
-    /// Getter accessor.
-    Getter,
-    /// Setter accessor.
-    Setter,
-    /// Enum member.
-    #[serde(rename = "enumMember")]
-    EnumMember,
-    /// TypeScript index signature.
-    #[serde(rename = "indexSignature")]
-    IndexSignature,
-}
-
-impl NormalizedMemberKind {
-    /// Converts extractor-level member kinds into public member kinds.
-    #[must_use]
-    pub fn from_doc_item_kind(kind: DocItemKind) -> Option<Self> {
-        match kind {
-            DocItemKind::Property => Some(Self::Property),
-            DocItemKind::Method => Some(Self::Method),
-            DocItemKind::Constructor => Some(Self::Constructor),
-            DocItemKind::Getter => Some(Self::Getter),
-            DocItemKind::Setter => Some(Self::Setter),
-            DocItemKind::EnumMember => Some(Self::EnumMember),
-            DocItemKind::IndexSignature => Some(Self::IndexSignature),
-            DocItemKind::Module
-            | DocItemKind::Function
-            | DocItemKind::Class
-            | DocItemKind::Interface
-            | DocItemKind::Type
-            | DocItemKind::Enum
-            | DocItemKind::Variable => None,
-        }
-    }
-
-    /// Returns the JavaScript-facing string for this kind.
-    #[must_use]
-    pub fn as_str(self) -> &'static str {
-        match self {
-            Self::Property => "property",
-            Self::Method => "method",
-            Self::Constructor => "constructor",
-            Self::Getter => "getter",
-            Self::Setter => "setter",
-            Self::EnumMember => "enumMember",
-            Self::IndexSignature => "indexSignature",
-        }
-    }
-}
-
-/// Normalized parameter documentation.
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct NormalizedParamDoc {
-    /// Parameter name.
-    pub name: String,
-    /// Parameter type, or `unknown` if it cannot be inferred.
-    pub type_annotation: String,
-    /// Parameter description.
-    pub description: String,
-    /// Whether the parameter is optional.
-    pub optional: bool,
-    /// Default value if specified.
-    pub default_value: Option<String>,
-}
-
-/// Normalized return documentation.
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct NormalizedReturnDoc {
-    /// Return type, or `unknown` if it cannot be inferred.
-    pub type_annotation: String,
-    /// Return value description.
-    pub description: String,
-    /// Members of an inline object literal return type.
-    #[serde(default)]
-    pub members: Vec<NormalizedMember>,
-}
-
-/// Normalized exception/error documentation.
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct NormalizedThrowsDoc {
-    /// Error type, when documented.
-    pub type_annotation: Option<String>,
-    /// Error condition description.
-    pub description: String,
-}
-
-/// Normalized type parameter documentation (`<T extends C = D>`).
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct NormalizedTypeParam {
-    /// Type parameter name (e.g. `T`).
-    pub name: String,
-    /// Constraint after `extends`, when present.
-    pub constraint: Option<String>,
-    /// Default type after `=`, when present.
-    pub default: Option<String>,
-    /// Description merged from a `@typeParam` / `@template` tag.
-    pub description: String,
-}
-
-/// Normalized documentation for a member of a class/interface/type/enum entry.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct NormalizedMember {
-    /// Member name.
-    pub name: String,
-    /// Member kind.
-    pub kind: NormalizedMemberKind,
-    /// Human-readable description.
-    pub description: String,
-    /// Full member signature, for callable members.
-    pub signature: Option<String>,
-    /// Property or enum member type/value annotation.
-    pub type_annotation: Option<String>,
-    /// Default value extracted from `@default` / `@defaultValue`.
-    #[serde(default)]
-    pub default_value: Option<String>,
-    /// Parameters, if any.
-    #[serde(default)]
-    pub params: Vec<NormalizedParamDoc>,
-    /// Member type parameters. Populated only when type-parameter docs are
-    /// enabled (opt-in); empty otherwise.
-    #[serde(default)]
-    pub type_parameters: Vec<NormalizedTypeParam>,
-    /// Return documentation, if any.
-    pub returns: Option<NormalizedReturnDoc>,
-    /// Exceptions/errors documented with `@throws` / `@exception`.
-    #[serde(default)]
-    pub throws: Vec<NormalizedThrowsDoc>,
-    /// Nested members owned by this member (for property-owned object literals).
-    #[serde(default)]
-    pub members: Vec<NormalizedMember>,
-    /// Whether the member is optional.
-    #[serde(default)]
-    pub optional: bool,
-    /// Whether the member is readonly.
-    #[serde(default)]
-    pub readonly: bool,
-    /// Whether the member is static.
-    #[serde(default)]
-    pub r#static: bool,
-    /// Whether the member is marked private.
-    #[serde(default)]
-    pub private: bool,
-    /// Custom JSDoc tags.
-    #[serde(default)]
-    pub tags: BTreeMap<String, String>,
-    /// Declaration start line.
-    pub line: u32,
-    /// Declaration end line.
-    pub end_line: u32,
-}
-
-impl Default for NormalizedMember {
-    fn default() -> Self {
-        Self {
-            name: String::new(),
-            kind: NormalizedMemberKind::default(),
-            description: String::new(),
-            signature: None,
-            type_annotation: None,
-            default_value: None,
-            params: Vec::new(),
-            type_parameters: Vec::new(),
-            returns: None,
-            throws: Vec::new(),
-            members: Vec::new(),
-            optional: false,
-            readonly: false,
-            r#static: false,
-            private: false,
-            tags: BTreeMap::new(),
-            line: 1,
-            end_line: 1,
-        }
-    }
-}
-
-/// Normalized documentation entry consumed by generated API docs.
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct NormalizedDocEntry {
-    /// Entry name.
-    pub name: String,
-    /// Entry kind.
-    pub kind: NormalizedDocKind,
-    /// Human-readable description.
-    pub description: String,
-    /// Parameters, if any.
-    pub params: Vec<NormalizedParamDoc>,
-    /// Return documentation, if any.
-    pub returns: Option<NormalizedReturnDoc>,
-    /// Exceptions/errors documented with `@throws` / `@exception`.
-    #[serde(default)]
-    pub throws: Vec<NormalizedThrowsDoc>,
-    /// Example blocks.
-    pub examples: Vec<String>,
-    /// Custom JSDoc tags.
-    pub tags: BTreeMap<String, String>,
-    /// Whether the entry is marked private.
-    pub private: bool,
-    /// Source file path.
-    pub file: String,
-    /// Declaration start line.
-    pub line: u32,
-    /// Declaration end line.
-    pub end_line: u32,
-    /// Signature text.
-    pub signature: Option<String>,
-    /// Extended base class/interface names.
-    #[serde(default)]
-    pub extends: Vec<String>,
-    /// Implemented interface names.
-    #[serde(default)]
-    pub implements: Vec<String>,
-    /// Whether a function declaration carries an implementation body (`false` for
-    /// overload signatures and ambient declarations). Used to hide the
-    /// implementation signature when grouping overloads on TypeDoc symbol pages.
-    #[serde(default)]
-    pub has_body: bool,
-    /// Members belonging to class/interface/type/enum entries.
-    #[serde(default)]
-    pub members: Vec<NormalizedMember>,
-    /// Declaration type parameters. Populated only when type-parameter docs are
-    /// enabled (opt-in); empty otherwise.
-    #[serde(default)]
-    pub type_parameters: Vec<NormalizedTypeParam>,
-}
-
-impl Default for NormalizedDocEntry {
-    fn default() -> Self {
-        Self {
-            name: String::new(),
-            kind: NormalizedDocKind::default(),
-            description: String::new(),
-            params: Vec::new(),
-            returns: None,
-            throws: Vec::new(),
-            examples: Vec::new(),
-            tags: BTreeMap::new(),
-            private: false,
-            file: String::new(),
-            line: 1,
-            end_line: 1,
-            signature: None,
-            extends: Vec::new(),
-            implements: Vec::new(),
-            has_body: false,
-            members: Vec::new(),
-            type_parameters: Vec::new(),
-        }
-    }
-}
 
 /// Normalizes extracted documentation items into API reference entries.
 #[must_use]
